+++ CHANGELOG.rst
@@ -21,6 +21,7 @@ Added
 ~~~~~
 
 - Add caching by default to PyJWKClient `#611 <https://github.com/jpadilla/pyjwt/pull/611>`__
+- Add support for ES256K algorithm `#629 <https://github.com/jpadilla/pyjwt/pull/629>`__
 
 `v2.0.1 <https://github.com/jpadilla/pyjwt/compare/2.0.0...2.0.1>`__
 --------------------------------------------------------------------